step11 b

+++ src/main/java/com/ll/domain/wiseSaying/controller/WiseSayingController.java
@@ -3,11 +3,9 @@ package com.ll.domain.wiseSaying.controller;
 import com.ll.AppContext;
 import com.ll.domain.wiseSaying.entity.WiseSaying;
 import com.ll.domain.wiseSaying.service.WiseSayingService;
+import com.ll.global.rq.Rq;
 
-import java.util.Arrays;
-import java.util.Map;
 import java.util.Scanner;
-import java.util.stream.Collectors;
 
 public class WiseSayingController {
     private final Scanner scanner;
@@ -38,18 +36,13 @@ public class WiseSayingController {
         }
     }
 
-    public void actionDelete(String cmd) {
-        String[] cmdBits = cmd.split("\\?", 2);
-        String queryString = cmdBits[1];
+    public void actionDelete(Rq rq) {
+        int id = rq.getParamsAsInt("id", -1);
 
-        Map<String, String> params = Arrays
-                .stream(queryString.split("&"))
-                .map(e -> e.split("=", 2))
-                .filter(e -> e.length == 2 && !e[0].isBlank() && !e[1].isBlank())
-                .collect(Collectors.toMap(e -> e[0].trim(), e -> e[1].trim()));
-
-        String idStr = params.getOrDefault("id", "");
-        int id = Integer.parseInt(idStr);
+        if (id == -1 ) {
+            System.out.println("id를 숫자로 입력해주세요.");
+            return;
+        }
 
         boolean isDeleted = wiseSayingService.delete(id);
 
@@ -61,18 +54,13 @@ public class WiseSayingController {
         System.out.printf("%d번 명언이 삭제되었습니다.\n", id);
     }
 
-    public void actionModify(String cmd) {
-        String[] cmdBits = cmd.split("\\?", 2);
-        String queryString = cmdBits[1];
+    public void actionModify(Rq rq) {
+        int id = rq.getParamsAsInt("id", -1);
 
-        Map<String, String> params = Arrays
-                .stream(queryString.split("&"))
-                .map(e -> e.split("=", 2))
-                .filter(e -> e.length == 2 && !e[0].isBlank() && !e[1].isBlank())
-                .collect(Collectors.toMap(e -> e[0].trim(), e -> e[1].trim()));
-
-        String idStr = params.getOrDefault("id", "");
-        int id = Integer.parseInt(idStr);
+        if (id == -1 ) {
+            System.out.println("id를 숫자로 입력해주세요.");
+            return;
+        }
 
         WiseSaying wiseSaying = wiseSayingService.findById(id);
